remove todo - test cross-platform
Theme verified to work on Chromium across major platforms - tested
on MacOS, Linux & Windows. WSL/2 was a PITA to get working due to
missing dependencies of puppeteer. I couldn't get the resume-cli
tool to output a PDF (fix out of scope) however, I could serve
the theme and Save-to-PDF via the browser print-page option.

For reference, this answer on Stack Overflow was an absolute gem
- https://stackoverflow.com/a/71283368/1238344